connected AdminPages to projec_java Main
