+++ src/adminOld/AdminPlaces.java
@@ -1,4 +1,4 @@
-package admin;
+package adminOld;
 
 import java.awt.BorderLayout;
 import java.awt.Color;
