+++ src/adminOld/PlacesDAO.java
@@ -1,4 +1,4 @@
-package admin;
+package adminOld;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
