+++ src/adminOld/UserVO.java
@@ -1,4 +1,4 @@
-package admin;
+package adminOld;
 
 public class UserVO {
 	private String userid, userpwd, username, userbirthday, useremail, userphone, userterms;
