+++ src/project_admin/ButtonColumn.java
@@ -1,4 +1,4 @@
-package admin2;
+package project_admin;
 
 import java.awt.*;
 import java.awt.event.*;
